Missing Licence / Improved documentation
Start to create an application which consist in verifying the compliance of IED contents and its "theoritical" data contents. Typically checking that initial value for Data Attributes are actually set in the IED (via Data Type definition and via DAI  instances). There will be two ways to this:  one is generating the Python to do such test, the other one will directly query the IED on line.